==> notebooks/00_main_nbk.ipynb <==
## Collecting all datasets

This notebook is used to fetch all the datasets in the sources

import pandas as pd
from logging import getLogger

from src.data_collection.comuni import fetch_dataset_comuni
from src.data_collection.indicators import get_indicators
from src.data_collection.indicators_sdg import get_indicators_and_sdg 
from src.data_collection.missioni import fetch_pnrr_missions
from src.data_collection.p01 import get_p01
from src.data_collection.p02 import get_p02
from src.data_collection.p03 import get_p03
from src.data_collection.p04 import get_p04
from src.data_collection.p05 import get_p05
from src.data_collection.targets import get_targets
from src.data_collection.localizzazioni import get_localizzazioni
from src.data_collection.regis import get_regis_indicators

p01 = get_p01()

print(p01.shape)
print(p01.columns)
p01.head()

# p01.to_parquet("data/updated/p01.parquet")

p02 = get_p02() # NOTE -> corretto che non esista più (fonte A.Borruso)

print(p02.shape)
print(p02.columns)
p02.head()

# p02.to_parquet("data/updated/p02.parquet")

p03 = get_p03()
print(p03.shape)
print(p03.columns)
p03.head()

# p03.to_parquet("data/updated/p03.parquet") 

p04 = get_p04()
print(p04.shape)
print(p04.columns)
p04.head()

# p04.to_parquet("data/updated/p04.parquet")

p05 = get_p05()
print(p05.shape)
print(p05.columns)
p05.head()

# p05.to_parquet("data/updated/p05.parquet")

comuni = fetch_dataset_comuni()
print(comuni.shape)
print(comuni.columns)
comuni.head()

# comuni.to_parquet("data/updated/comuni.parquet")

missions, components = fetch_pnrr_missions()
print(missions.shape)
print(missions.columns)
missions.head()

# missions.to_parquet("data/updated/missions.parquet")

print(components.shape)
print(components.columns)
components.head()

# components.to_parquet("data/updated/components.parquet")

indicators = get_indicators()
print(indicators.shape)
print(indicators.columns)
indicators.head()

# indicators.to_parquet("data/updated/indicators.parquet")

indicators_sdg = get_indicators_and_sdg()
print(indicators_sdg.shape)
print(indicators_sdg.columns)
indicators_sdg.head()

# indicators_sdg.to_parquet("data/updated/indicators_sdg.parquet")

targets = get_targets()
print(targets.shape)
print(targets.columns)
targets.head()

# targets.to_parquet("data/updated/targets.parquet")

regis_comm_ind = get_regis_indicators()
print(regis_comm_ind.shape)
print(regis_comm_ind.columns)
regis_comm_ind.head()

# regis_comm_ind.to_parquet("data/updated/universo_regis.parquet")
