==> notebooks/02_master_dati_italia_domani.ipynb <==
# Merge Master Table 
Notebook di merging della master table (ottenuta nel Notebook 1) con i dataset provenienti da Italia Domani 

import pandas as pd
pd.set_option('display.max_columns', None)
master_table = pd.read_csv("data/updated/master_table_base.csv", sep=";")

# Indicatori

indicators = pd.read_parquet("data/updated/indicators.parquet")

master_table_ID = master_table.merge(
    indicators,
    how='left',
    left_on='Codice Univoco Submisura',
    right_on='Codice Univoco Submisura'
)
master_table_ID = master_table_ID.drop_duplicates()
print(master_table_ID.shape)

master_table_ID.columns

colonne = [
    'CIG', 
    'CUP', 
    'Regione', 
    'provincia', 
    'Comune',
    'importo_complessivo_gara', 
    'CLASSE_IMPORTO', 
    'Missione_x', 
    'Descrizione Missione_x',
    'Componente_x',
    'cod_mis_premiale', 
    'misura_premiale',
    'CATEGORIA', 
    'flag_quote', 
    'quota_femminile', 
    'quota_giovanile',
    'cod_mot_deroga', 
    'mot_deroga', 
    'data_pubblicazione',
    'anno_pubblicazione', 
    'FLAG_URGENZA', 
    'MOTIVO_URGENZA', 
    'ESITO',
    'Codice Univoco Submisura', 
    'Programma',
    'Descrizione Componente',
    'ID Misura', 
    'Codice Univoco Misura', 
    'Descrizione Misura',
    'ID Submisura', 
    'Codice CID', 
    'Descrizione Submisura',
    'Amministrazione Titolare', 
    'Codice Indicatore',
    'Descrizione Indicatore', 
    'Unità di misura', 
    'Stock/Flusso'
]

master_table_ID = master_table_ID[colonne]

master_table_ID.head()

master_table_ID.to_parquet('data/helpers/master_ID1.parquet')

## Indicatori SDG

import pandas as pd
pd.set_option('display.max_columns', None)

master_table_ID = pd.read_parquet('data/helpers/master_ID1.parquet')
indicatori_sdg = pd.read_parquet("data/updated/indicators_sdg.parquet")

master_table_ID = master_table_ID.merge(
    indicatori_sdg,
    how='left',
    left_on='Codice Univoco Submisura',
    right_on='COD_SUB_MISURA'
)
master_table_ID = master_table_ID.drop_duplicates()
print(master_table_ID.shape)

colonne = [
    'CIG', 
    'CUP', 
    'Regione', 
    'provincia', 
    'Comune',
    'importo_complessivo_gara', 
    'CLASSE_IMPORTO', 
    'Missione_x', 
    'Descrizione Missione_x',
    'Componente_x',
    'DES_COMPONENTE',
    'cod_mis_premiale', 
    'misura_premiale',
    'CATEGORIA', 
    'flag_quote', 
    'quota_femminile', 
    'quota_giovanile',
    'cod_mot_deroga', 
    'mot_deroga', 
    'data_pubblicazione',
    'anno_pubblicazione', 
    'FLAG_URGENZA', 
    'MOTIVO_URGENZA', 
    'ESITO',
    'Codice Univoco Submisura', 
    'Programma',
    'COD_MISURA',
    'Codice Univoco Misura', 
    'DES_MISURA',
    'DES_LIVELLO',
    'ID Submisura', 
    'Codice CID', 
    'Descrizione Submisura',
    'Amministrazione Titolare', 
    'Codice Indicatore',
    'Descrizione Indicatore', 
    'Unità di misura', 
    'Stock/Flusso',
    'IMPORTO',
    'DES_INVESTIMENTO',
    'DES_PRESTITO',
    'DES_AMMINISTRAZIONE',
    'PREVALENTE',
    'INDICATORE',
    'DES_INDICATORE',
    'DES_INDAGINE',
    'NOTA',
    'DES_FONTE',
    'DES_UNITA',
    'GOAL',
    'DES_GOAL',
    'TARGET',
    'DES_TARGET',
    'INDICATOR',
    'DES_INDICATOR'
]

master_table_ID = master_table_ID[colonne]
master_table_ID.shape

master_table_ID.head()

master_table_ID.to_parquet('data/helpers/master_ID2.parquet')

## Targets

import pandas as pd
pd.set_option('display.max_columns', None)

master_table_ID = pd.read_parquet('data/helpers/master_ID2.parquet')
targets = pd.read_parquet("data/updated/targets.parquet")

master_table_ID = master_table_ID.merge(
    targets,
    how='left',
    left_on='CUP',
    right_on='CUP'
)
master_table_ID = master_table_ID.drop_duplicates()
print(master_table_ID.shape)

colonne = [
    'CIG', 
    'CUP', 
    'Regione', 
    'provincia', 
    'Comune',
    'importo_complessivo_gara', 
    'CLASSE_IMPORTO', 
    'Missione_x', 
    'Descrizione Missione_x',
    'Componente_x',
    'DES_COMPONENTE',
    'cod_mis_premiale', 
    'misura_premiale',
    'CATEGORIA', 
    'flag_quote', 
    'quota_femminile', 
    'quota_giovanile',
    'cod_mot_deroga', 
    'mot_deroga', 
    'data_pubblicazione',
    'anno_pubblicazione', 
    'FLAG_URGENZA', 
    'MOTIVO_URGENZA', 
    'ESITO',
    'Codice Univoco Submisura_x', 
    'Programma',
    'COD_MISURA',
    'Codice Univoco Misura', 
    'DES_MISURA',
    'DES_LIVELLO',
    'ID Submisura', 
    'Codice CID', 
    'Descrizione Submisura_x',
    'Amministrazione Titolare', 
    'Codice Indicatore_x',
    'Descrizione Indicatore_x', 
    'Unità di misura', 
    'Descrizione Unità di Misura',
    'Stock/Flusso',
    'IMPORTO',
    'DES_INVESTIMENTO',
    'DES_PRESTITO',
    'DES_AMMINISTRAZIONE',
    'PREVALENTE',
    'INDICATORE',
    'DES_INDICATORE',
    'DES_INDAGINE',
    'NOTA',
    'DES_FONTE',
    'DES_UNITA',
    'GOAL',
    'DES_GOAL',
    'TARGET',
    'DES_TARGET',
    'INDICATOR',
    'DES_INDICATOR',
    'Codice Locale Progetto',
    'Valore Programmato',
    'Valore Realizzato',
    'Mese',
    'Anno',
    'Data di Aggiornamento'
]

master_table_ID = master_table_ID[colonne]
master_table_ID.shape

master_table_ID.to_parquet('data/helpers/master_ID3.parquet')

## Universo Regis 

import pandas as pd
pd.set_option('display.max_columns', None)

master_table_ID = pd.read_parquet('data/helpers/master_ID3.parquet')
indicatori_com_regis = pd.read_parquet("data/updated/universo_regis.parquet")

master_table_ID.shape

master_table_ID.columns

master_table_ID['Codice Indicatore_x']

master_table_ID['Anno'].isna().sum()

master_table_ID['CUP'].nunique()

indicatori_com_regis['CUP'].nunique()

indicatori_com_regis.shape

indicatori_com_regis.columns

indicatori_com_regis['FLG_INDICATORI_GENERE']

indicatori_com_regis['FLG_INDICATORI_GENERE'].nunique()

indicatori_com_regis['FLG_INDICATORI_GENERE'].isna().sum()

indicatori_com_regis.dropna

cols_to_drop = [
    'Codice Univoco Submisura', 
    'Descrizione Submisura', 
    'Codice Indicatore', 
    'Descrizione Indicatore',
    'Unità di Misura', 
    'Descrizione Unità di Misura',
    'Descrizione Tipo Indicatore', 
    'Mese', 
    'Anno', 
    'Data di Estrazione'
]

indicatori_com_regis.drop(columns=cols_to_drop, inplace=True)
indicatori_com_regis.shape

master_table_ID = master_table_ID.merge(
    indicatori_com_regis,
    how='left',
    left_on='CUP',
    right_on='CUP'
)
master_table_ID = master_table_ID.drop_duplicates()
print(master_table_ID.shape)

master_table_ID.to_parquet('data/updated/master_ID4.parquet')
