==> notebooks/03_duckdb_data.ipynb <==
import duckdb
import pandas as pd

## Database Views 
Per ottenere il DB con le varie viste dei dati PNRR, seguire [questo](https://aborruso.github.io/posts/tutto-il-pnrr-in-500k/) link. Posizionare il DB scaricato nella cartella principale per eseguire query direttamente con DuckDB


db_path = 'pnrr.db' 

con = duckdb.connect(database=db_path, read_only=True)


# List tables in the database (optional, for exploration)
print("Tables in the database:")
print(con.execute("SHOW TABLES;").fetchdf())

df = con.execute(f"SELECT * FROM {'info_viste'};").fetchdf()

df.shape

df.columns

for _, row in df.iterrows():
    print(f"View: {row["view"]}")
    print(f"Title: {row['title']}")
    print(f"URL: {row['page_url']} \n\n")
